# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: AsterFerry Controller API
  version: 1.0.0
  description: Authoritative control-plane API for the AsterFerry v1.0.0 normalized architecture. Gateway and Agent business traffic never traverses this API.
servers:
  - url: /api/v1
...
paths:
  /controller/update:
    get:
      tags: [operations]
      operationId: getControllerUpdate
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses: { '200': { description: Stable Controller release status, content: { application/json: { schema: { $ref: '#/components/schemas/ControllerUpdateResponse' } } } } }
...
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } }
      responses:
        '200': { description: Authenticated, content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRF' }]
      responses: { '204': { description: Logged out }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /me:
    get:
      tags: [auth]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses: { '200': { description: Current user, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } } }
  /nodes:
    get:
      tags: [nodes]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ name: kind, in: query, description: Filter by configured Node behavior, schema: { type: string, enum: [gateway, agent] } }]
      responses: { '200': { description: Nodes, content: { application/json: { schema: { $ref: '#/components/schemas/NodeList' } } } } }
    post:
      tags: [nodes]
      operationId: createNode
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CSRF' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/NodeInput' } } }
      responses: { '201': { description: Created }, '409': { $ref: '#/components/responses/Conflict' } }
...
  /node-installations:
    get:
      tags: [operations]
      description: List installation intents that have not enrolled yet. The response never contains the one-time token.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200':
          description: Pending node installations.
          content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/PendingNodeInstallation' } } } } } }
    post:
      tags: [operations]
      operationId: createNodeInstallation
      description: Create a generic Node installation intent and return a one-time platform installer command. The enrolled identity is created when the installer completes Enroll; Gateway or Agent behavior is configured later through the node spec endpoint.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CSRF' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/NodeInstallationRequest' } } }
      responses:
        '201':
          description: One-time installer command generated.
          content: { application/json: { schema: { $ref: '#/components/schemas/NodeBootstrapResponse' } } }
        '409': { description: Pending installation conflict or unrecoverable token retry, content: { application/json: { schema: { $ref: '#/components/schemas/SecretAlreadyCreated' } } } }
        '503': { description: Controller bootstrap configuration is unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  /services:
    get: { tags: [resources], security: [{ bearerAuth: [] }, { cookieAuth: [] }], responses: { '200': { description: Services } } }
    post:
      tags: [resources]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CSRF' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Service' } } } }
      responses: { '201': { description: Created }, '409': { $ref: '#/components/responses/Conflict' } }
...
  /assignments:
    get: { tags: [resources], security: [{ bearerAuth: [] }, { cookieAuth: [] }], responses: { '200': { description: Assignments } } }
    post:
      tags: [resources]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CSRF' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Assignment' } } } }
      responses: { '201': { description: Created }, '409': { $ref: '#/components/responses/Conflict' } }
...
  /enrollment-tokens:
    get: { tags: [operations], security: [{ bearerAuth: [] }, { cookieAuth: [] }], responses: { '200': { description: Enrollment tokens without plaintext secrets } } }
    post:
      tags: [operations]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CSRF' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/EnrollmentTokenRequest' } } } }
      responses: { '201': { description: Plaintext is returned once }, '409': { description: Request was already created; plaintext cannot be recovered, content: { application/json: { schema: { $ref: '#/components/schemas/SecretAlreadyCreated' } } } } }
...
  /users:
    get: { tags: [auth], security: [{ bearerAuth: [] }, { cookieAuth: [] }], responses: { '200': { description: Users } } }
    post:
      tags: [auth]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CSRF' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserInput' } } } }
      responses: { '201': { description: Created } }
...
  /runtime/settings:
    get:
      tags: [runtime]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses: { '200': { description: Runtime operation settings } }
    put:
      tags: [runtime]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CSRF' }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [advanced_operations_enabled], properties: { advanced_operations_enabled: { type: boolean } } } } }
      responses: { '200': { description: Runtime operation settings updated } }
  /runtime/connections:
    get:
      tags: [runtime]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters:
        - { name: node_id, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: source_ip, in: query, schema: { type: string, description: IPv4 or IPv6 address } }
        - { name: assignment_id, in: query, schema: { type: string } }
        - { name: service_id, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000 } }
      responses: { '200': { description: Current and recently closed connection metadata } }
  /runtime/events:
    get:
      tags: [runtime]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ name: node_id, in: query, schema: { type: string } }, { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000 } }]
      responses: { '200': { description: Runtime lifecycle events retained for 30 days } }
  /runtime/traffic:
    get:
      tags: [runtime]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ name: node_id, in: query, schema: { type: string } }, { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 2000 } }]
      responses: { '200': { description: Per-minute traffic rollups } }
  /runtime/stream:
    get:
      tags: [runtime]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ name: node_id, in: query, schema: { type: string } }]
      responses: { '200': { description: Server-sent runtime metadata change notifications, content: { text/event-stream: {} } } }
...
  /audit:
    get:
      tags: [audit]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000, default: 100 } }]
      responses: { '200': { description: Audit records } }
  /events:
    get:
      tags: [audit]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses: { '200': { description: Event records } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    cookieAuth: { type: apiKey, in: cookie, name: af_session }
    nodeEnrollmentAuth: { type: apiKey, in: header, name: X-AsterFerry-Enrollment-Token, description: One-time token returned by the Dashboard node installation command generator. }
  parameters:
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: false, schema: { type: string, maxLength: 128 } }
    CSRF: { name: X-CSRF-Token, in: header, required: false, schema: { type: string, minLength: 32 } }
  responses:
    Unauthorized: { description: Authentication required }
    Conflict: { description: Revision or unique binding conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  schemas:
...
    ControllerUpdateState:
      type: object
      required: [schema_version, current_version, channel, deployment, supported, state, updated_at]
      properties:
        schema_version: { type: integer }
        current_version: { type: string }
        channel: { type: string, enum: [stable] }
        deployment: { type: string }
        supported: { type: boolean }
        state: { type: string, enum: [disabled, checking, up_to_date, available, applying, failed, rolled_back, manual_required, unsupported] }
        latest_version: { type: string }
        latest_url: { type: string, format: uri }
        published_at: { type: string, format: date-time }
        last_checked_at: { type: string, format: date-time }
        last_error: { type: string }
        target_version: { type: string }
        previous_version: { type: string }
        updated_at: { type: string, format: date-time }
    ControllerUpdateResponse:
      type: object
      required: [reason, status]
      properties: { reason: { type: string }, status: { $ref: '#/components/schemas/ControllerUpdateState' } }
...
    LoginRequest:
      type: object
      required: [username, password]
      properties: { username: { type: string }, password: { type: string, format: password } }
    LoginResponse:
      type: object
      properties: { user: { $ref: '#/components/schemas/User' }, csrf_token: { type: string } }
    User:
      type: object
      properties: { id: { type: string }, username: { type: string }, role: { type: string, enum: [viewer, operator, admin] }, enabled: { type: boolean }, revision: { type: integer, format: int64 }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } }
    UserInput:
      type: object
      properties: { username: { type: string }, password: { type: string, format: password }, role: { type: string, enum: [viewer, operator, admin] } }
    NodeInput:
      type: object
      required: [id, name]
      properties: { id: { type: string }, name: { type: string }, labels: { type: object, additionalProperties: { type: string } }, enabled: { type: boolean } }
...
    NodeBootstrapResponse:
      type: object
      required: [node_id, platform, arch, version, script_source, installer_url, expires_at, command, state]
      properties:
        installation_id: { type: string, description: Pending installation identifier; set only by the install-first endpoint. }
        state: { type: string, enum: [install, replacement, pending], description: install for a new identity, replacement for a decommissioned identity being re-enrolled, and pending for an install-first intent that has not enrolled yet. }
        node_id: { type: string }
        platform: { type: string, enum: [linux, windows] }
        arch: { type: string, enum: [amd64, arm64] }
        version: { type: string, pattern: '^[0-9]+\.[0-9]+\.[0-9]+(-rc\.[0-9]+)?$' }
        script_source: { type: string, enum: [controller, github] }
        installer_url: { type: string, format: uri, description: Exact installer script URL used by the generated command. }
        expires_at: { type: string, format: date-time }
        command: { type: string }
    NodeInstallationRequest:
      type: object
      required: [name, platform, arch, script_source]
      description: Creates a pending installation intent. The Controller generates node_id when it is omitted. It does not create an enrolled node until the installer completes Enroll.
      properties:
        node_id: { type: string, minLength: 1, maxLength: 128, description: Optional immutable identity override; omit this field to let the Controller generate a unique ID. }
        name: { type: string, minLength: 1, maxLength: 256 }
        labels: { type: object, additionalProperties: { type: string } }
        enabled: { type: boolean, default: true }
        platform: { type: string, enum: [linux, windows] }
        arch: { type: string, enum: [amd64, arm64] }
        script_source: { type: string, enum: [controller, github], default: controller, description: Where the one-line installer script is downloaded from. The Controller option is recommended. }
    PendingNodeInstallation:
      type: object
      required: [node_id, name, enabled, platform, arch, script_source, expires_at, created_at]
      properties:
        node_id: { type: string }
        name: { type: string }
        labels: { type: object, additionalProperties: { type: string } }
        enabled: { type: boolean }
        platform: { type: string, enum: [linux, windows] }
        arch: { type: string, enum: [amd64, arm64] }
        script_source: { type: string, enum: [controller, github] }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
...
    Node:
      allOf:
        - $ref: '#/components/schemas/NodeInput'
        - type: object
          properties: { spec_kind: { type: string, enum: [gateway, agent], description: Configured behavior projection; omitted until a spec is saved. }, certificate_state: { type: string, enum: [pending, active, revoked, expired, decommissioned] }, certificate_serial: { type: string }, revision: { type: integer, format: int64 }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } }
    NodeList: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Node' } } } }
...
    Service: { type: object, required: [id, agent_id, protocol, local_target, public_bind], properties: { id: { type: string }, agent_id: { type: string }, protocol: { type: string, enum: [tcp, udp] }, local_target: { type: string }, public_bind: { type: string }, public_port: { type: integer, minimum: 0, maximum: 65535 }, gateway_selector: { type: object }, enabled: { type: boolean }, revision: { type: integer, format: int64 }, updated_at: { type: string, format: date-time } } }
    Assignment: { type: object, required: [id, gateway_id, agent_id, generation], properties: { id: { type: string }, gateway_id: { type: string }, agent_id: { type: string }, service_ids: { type: array, items: { type: string } }, bindings: { type: array, items: { type: object } }, generation: { type: integer, format: int64 }, state: { type: string }, public_endpoint: { type: string }, revision: { type: integer, format: int64 }, updated_at: { type: string, format: date-time } } }
    EnrollmentTokenRequest: { type: object, properties: { ttl_seconds: { type: integer, minimum: 1, maximum: 900 } } }
...
    SecretAlreadyCreated:
      type: object
      required: [error, token_recoverable]
      properties:
        error: { type: object, properties: { code: { type: string, enum: [already_created] }, message: { type: string } } }
        token_recoverable: { type: boolean, enum: [false] }
        metadata: { type: object }
        token_metadata: { type: object }
    ErrorDetail: { type: object, properties: { code: { type: string }, message: { type: string }, path: { type: string }, retryable: { type: boolean } } }
    ErrorResponse: { type: object, properties: { error: { $ref: '#/components/schemas/ErrorDetail' } } }